requirements.txt
openai>=1.0.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
aiohttp>=3.8.5
requests>=2.31.0
moviepy>=1.0.3
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
opencv-python>=4.5.5
numpy>=1.21.0
ffmpeg-python>=0.2.0 